Sets docker up and running without data
This commit sets up the requirements.txt and a basic dockerfile that
works to install the necessary requirements for tensorflow running on a
cpu. Will need access to a linux box with a gpu to test the ability to
run it on a GPU and with tensorflow-gpu, as the nvidia-docker runtime
does not work unless there is an nvidia card available.