Gets kfold validation working
Run kfold.py and pass the input data file and output directories and it
will do a 5 way k fold with a train: 0.7 val: 0.1 test: 0.2 split.
Updates run_evaluation_onTIF.py to not put a png in the root directory,
but rather put it in the prediction directory that you specified
earlier.