Integrates L3 data loader with prediction
This is the first commit that fully works making predictions from a list
of StudyImages sagittal MIPs. These MIPs are currently generated by iterating over
directories in Tim's script's format. It uses the same method to make
L3 slice predictions that is from the ct-slice-detection from the
reference paper. Right now, it outputs a sample of the predictions as
plots in Matplotlib, though this should be configurable in the future
(works nice in Pycharm because plots are non-blocking and are shown on
the side...).

Next steps are to find the L3 dcm corresponding to the prediction then
pass it to the muscle segmenter!

This commit also refactors the L3_finder.py a little bit, to make the
steps done on the StudyImages list a pipeline, which should be extended
with a timer that displays how long each step takes and which step is
being run. Definitely a nice to have but will help in the future.