Adds batch_to_ndarray method to utils
This method takes an iterable and batches its items into an Ndarray.
Will potentially be useful to speed up prediction, as the keras models
can take ndarrays, but right now are only taking one image at a time.
Will potentially need to use this to speed up processing of the 10k
images.