Adds programs to calculate segmentation area
The calculate_segmentation_area_for_dir.py script will calculate the
segmentation area for the subjects in a directory. It requires the path
to a directory with the original dicom images (in Tim's output format)
to get the spacing and resolution of the original images. It prints the
output in csv format.

The calculate_seg_areas_for_dirs.py script will accept a list of dirs
and call the previous script on each of the directories, outputing the
results to multiple csv files (for now, try excel).